move simplify_valid to symbolic_flat
will be using this for real_stride
